pyproject.toml
[tool.poetry]
name = "student-bluebook"
version = "0.4.0"
description = "Blue Book is an application that uses AI to generate multiple-choice questions to prepare for CompTIA Security+ exam"
authors = ["Ilya Smut <ilya.smut.off.g@gmail.com>"]
readme = "README.md"
packages = [{include = "bluebook"}]
homepage = "https://github.com/ilya-smut/blue-book"
repository = "https://github.com/ilya-smut/blue-book"
documentation = "https://github.com/ilya-smut/blue-book/wiki"

[tool.poetry.dependencies]
python = ">=3.10"
google-genai = ">=1.2.0,<2.0.0"
flask = ">=3.1.0,<4.0.0"
click = ">=8.1.8,<9.0.0"
pydantic = "^2.10.6"
sqlmodel = "^0.0.24"
bleach = "^6.2.0"

[tool.poetry.scripts]
bluebook = "bluebook.fl:bluebook"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"